requirements.txt
pylint==1.7.4
GitPython==2.1.6
requests==2.18.4
PyYAML==3.12
psycopg2==2.7.3.2
glob2==0.5
matplotlib==2.1.0
